requirements.txt
asgiref==3.2.7
Django==3.0.7
flake8==3.8.3
importlib-metadata==1.6.1
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
gunicorn==20.0.4
pymorphy2==0.8
whitenoise==5.1.0
pillow==7.2.0
phonenumbers==8.12.6
django-autocomplete-light==3.5.1
django-cached-modelforms==0.2.3
django-debug-toolbar==2.2
django-configurations==2.2
psycopg2==2.8.5
django-heroku==0.3.1
